MultiEconomy 2.0.0 support, EconomyProvider::getMoney() now returns float

+++ src/DaPigGuy/libPiggyEconomy/providers/EconomyProvider.php
@@ -30,9 +30,9 @@ abstract class EconomyProvider
 
     /**
      * @param Player $player
-     * @return int
+     * @return float
      */
-    abstract function getMoney(Player $player): int;
+    abstract function getMoney(Player $player): float;
 
     /**
      * @param Player $player
